+++ packages/std/src/errors/mod.rs
@@ -4,7 +4,7 @@ mod system_error;
 
 pub use std_error::{StdError, StdResult};
 pub use std_error_helpers::{
-    generic_err, invalid_base64, invalid_utf8, not_found, null_pointer, parse_err, serialize_err,
-    unauthorized, underflow,
+    generic_err, invalid_base64, invalid_utf8, not_found, parse_err, serialize_err, unauthorized,
+    underflow,
 };
 pub use system_error::{SystemError, SystemResult};