+++ packages/std/src/lib.rs
@@ -16,8 +16,8 @@ mod types;
 pub use crate::coins::{coin, coins, has_coins, Coin};
 pub use crate::encoding::Binary;
 pub use crate::errors::{
-    generic_err, invalid_base64, invalid_utf8, not_found, null_pointer, parse_err, serialize_err,
-    unauthorized, underflow, StdError, StdResult, SystemError, SystemResult,
+    generic_err, invalid_base64, invalid_utf8, not_found, parse_err, serialize_err, unauthorized,
+    underflow, StdError, StdResult, SystemError, SystemResult,
 };
 pub use crate::init_handle::{
     log, BankMsg, CosmosMsg, HandleResponse, HandleResult, InitResponse, InitResult, LogAttribute,